Refactor pin_lib to use Intel XED for static decoding. Split the decoder into decoder/x86_decoder where x86_decoder performs static decoding of instructions and decoder performs extraction of dynamic instruction information (using PIN). Similarly, gather_scatter is split into separate files for dynamic and static decoding. In preparation of supporting additional frontends that need to resuse the static instruction decoding.
