+++ src/pin/pin_lib/gather_scatter_addresses.h
@@ -24,12 +24,15 @@
 
 #include <unordered_map>
 #include <vector>
+#include <string>
+#include <ostream>
 #undef UNUSED   // there is a name conflict between PIN and Scarab
 #undef WARNING  // there is a name conflict between PIN and Scarab
-#include "pin.H"
+//#include "pin.H"
 #undef UNUSED   // there is a name conflict between PIN and Scarab
 #undef WARNING  // there is a name conflict between PIN and Scarab
 #include "../../ctype_pin_inst.h"
+#include "pin_api_to_xed.h"
 
 class gather_scatter_info {
  public:
@@ -50,65 +53,63 @@ class gather_scatter_info {
   gather_scatter_info::mask_reg_type get_mask_reg_type() const;
   bool                               data_dest_reg_set() const;
 
-  void   set_data_reg_total_width(const REG pin_reg);
-  void   set_data_lane_width_bytes(const UINT32 st_lane_width);
-  void   set_mask_reg(const REG pin_reg);
-  void   set_base_reg(const REG pin_reg);
-  void   set_index_reg(const REG pin_reg);
-  void   set_displacement(const ADDRDELTA displacement);
-  void   set_scale(const UINT32 scale);
-  void   set_index_lane_width_bytes(const UINT32 idx_lane_width);
+  void   set_data_reg_total_width(const xed_reg_enum_t pin_reg);
+  void   set_data_lane_width_bytes(const uint32_t st_lane_width);
+  uint32_t get_data_lane_width_bytes() const;
+  void   set_mask_reg(const xed_reg_enum_t pin_reg);
+  void   set_base_reg(const xed_reg_enum_t pin_reg);
+  void   set_index_reg(const xed_reg_enum_t pin_reg);
+  xed_reg_enum_t get_mask_reg() const;
+  xed_reg_enum_t get_base_reg() const;
+  xed_reg_enum_t get_index_reg() const;
+  void   set_displacement(const uint64_t displacement);
+  uint64_t get_displacement() const;
+  void   set_scale(const uint32_t scale);
+  uint32_t get_scale() const;
+  void   set_index_lane_width_bytes(const uint32_t idx_lane_width);
+  uint32_t get_index_lane_width_bytes() const;
   void   compute_num_mem_ops();
-  UINT32 get_num_mem_ops() const;
+  uint32_t get_num_mem_ops() const;
   void   verify_fields_for_mem_access_info_generation() const;
-  vector<PIN_MEM_ACCESS_INFO> compute_mem_access_infos(
-    const CONTEXT* ctxt) const;
   bool base_reg_is_gr32() const;
 
-  friend ostream& operator<<(ostream& os, const gather_scatter_info& sinfo);
+  friend std::ostream& operator<<(std::ostream& os, const gather_scatter_info& sinfo);
 
  private:
   gather_scatter_info::type          _type;
   gather_scatter_info::mask_reg_type _mask_reg_type;
-  UINT32                             _data_vector_reg_total_width_bytes;
-  UINT32                             _data_lane_width_bytes;
-  REG                                _mask_reg;
-  REG                                _base_reg;
-  REG                                _index_reg;
-  ADDRDELTA                          _displacement;
-  UINT32                             _scale;
-  UINT32                             _index_lane_width_bytes;
-  UINT32                             _num_mem_ops;
+  uint32_t                             _data_vector_reg_total_width_bytes;
+  uint32_t                             _data_lane_width_bytes;
+  xed_reg_enum_t                                _mask_reg;
+  xed_reg_enum_t                                _base_reg;
+  xed_reg_enum_t                                _index_reg;
+  uint64_t                          _displacement;
+  uint32_t                             _scale;
+  uint32_t                             _index_lane_width_bytes;
+  uint32_t                             _num_mem_ops;
 
-  bool      is_non_zero_and_powerof2(const UINT32 v) const;
-  UINT32    pin_xyzmm_reg_width_in_bytes(const REG pin_xyzmm_reg) const;
-  ADDRDELTA compute_base_reg_addr_contribution(const CONTEXT* ctxt) const;
-  ADDRDELTA compute_base_index_addr_contribution(
-    const PIN_REGISTER& vector_index_reg_val, const UINT32 lane_id) const;
-  PIN_MEMOP_ENUM type_to_PIN_MEMOP_ENUM() const;
+  bool      is_non_zero_and_powerof2(const uint32_t v) const;
+  uint32_t    pin_xyzmm_reg_width_in_bytes(const xed_reg_enum_t pin_xyzmm_reg) const;
   void           verify_mask_reg() const;
-  bool           extract_mask_on(const PIN_REGISTER& mask_reg_val_buf,
-                                 const UINT32        lane_id) const;
 };
 
-void add_to_gather_scatter_info_storage(const ADDRINT iaddr,
-                                        const bool    is_gather,
-                                        const bool    is_scatter,
-                                        const INT32   category);
-void set_gather_scatter_reg_operand_info(const ADDRINT iaddr, const REG pin_reg,
+typedef std::unordered_map<ADDRINT, gather_scatter_info> scatter_info_map;
+gather_scatter_info add_to_gather_scatter_info_storage(const ADDRINT iaddr,
+                                                       const bool    is_gather,
+                                                       const bool    is_scatter,
+                                            const xed_category_enum_t category);
+void set_gather_scatter_reg_operand_info(const ADDRINT iaddr,
+                                         const xed_reg_enum_t pin_reg,
                                          const bool operandRead,
                                          const bool operandWritten);
 void set_gather_scatter_memory_operand_info(
-  const ADDRINT iaddr, const REG pin_base_reg, const REG pin_index_reg,
-  const ADDRDELTA displacement, const UINT32 scale, const bool operandReadOnly,
-  const bool operandWritenOnly);
+                                            const ADDRINT iaddr,
+                                            const xed_reg_enum_t pin_base_reg,
+                                            const xed_reg_enum_t pin_index_reg,
+                                            const uint32_t scale,
+                                            const bool operandReadOnly,
+                                            const bool operandWritenOnly);
 void finalize_scatter_info(const ADDRINT iaddr, ctype_pin_inst* info);
-vector<PIN_MEM_ACCESS_INFO>
-  get_gather_scatter_mem_access_infos_from_gather_scatter_info(
-    const CONTEXT* ctxt, const PIN_MULTI_MEM_ACCESS_INFO* infos_from_pin);
-void update_gather_scatter_num_ld_or_st(const ADDRINT                   iaddr,
-                                        const gather_scatter_info::type type,
-                                        const uint      num_maskon_memops,
-                                        ctype_pin_inst* info);
+void init_reg_xed_to_pin_map();
 
 #endif  // __SCATTER_H__
